npc_ric-main/generator/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
docxtpl==0.16.0
python-dotenv==1.0.0
fastapi==0.111.0
uvicorn==0.29.0
